bugfix: Fix the space bar issue at its root:
With the upgrade to React 19, there is a slight conflict with Monaco Editor. Because Monaco makes heavy use of useRef and React 19 made some significant changes to the behavior of useRef, there are a number of knock-on effects.

I am using Monaco's internal keyboard action API to bypass issues with DOM event keybinding

Signed-off-by: Jared Scott <jared.scott@datarecce.io>